requirements.txt
google-api-python-client>=2.100
pandas>=2.0
vaderSentiment>=3.3
python-dotenv>=1.0
googletrans==4.0.0rc1
requests>=2.31
customtkinter>=5.2
openpyxl>=3.1
tqdm>=4.66